requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
numpy==1.26.4
reportlab==4.1.0
google-generativeai
python-dotenv
Flask-Caching
Flask-Limiter
Flask-WTF
